Completed fix for #243: now also with cases having parameters or requiring fixtures. Made the test only valid in python 3.
